language changing

+++ handlers/users/main_handler.py
@@ -1,4 +1,5 @@
 
+import logging
 from aiogram.types import Message, CallbackQuery
 from aiogram.dispatcher import FSMContext
 from keyboards.default.def_keyboards import def_keyboard, back_button
@@ -13,7 +14,7 @@ nar_key=narrow()
 
 @dp.message_handler(text="Menu")
 async def bot_start(message: Message, state:FSMContext):
-    await message.answer("Bizdagi bosh menu 🥡:", reply_markup=await def_keyboard())
+    await message.answer("The main menu in our fast food kitchen 🥡:", reply_markup=await def_keyboard())
     await state.set_state("menu")
 
 
@@ -21,8 +22,8 @@ async def bot_start(message: Message, state:FSMContext):
 
 @dp.message_handler(state="menu")
 async def select_category(message: Message, state:FSMContext):
-    category = message.text # chesen category
-    await message.answer("Nima tanlaysiz?", reply_markup=await def_keyboard(category)) # this func generates markup for carresponding category of item
+    category = message.text # chosen category
+    await message.answer("What do you choose ?", reply_markup=await def_keyboard(category)) # this func generates markup for corresponding category of items
     await state.set_state("select_item")
     await state.update_data({
         "category": category
@@ -49,6 +50,9 @@ async def select_category(message: Message, state:FSMContext):
     })
 
 
+#-------------------this part is responsible for callback_queries--------------------------------
+
+
 @dp.callback_query_handler(text="+", state="order")
 async def plus_handler(query:CallbackQuery, state: FSMContext):
     
@@ -60,8 +64,8 @@ async def plus_handler(query:CallbackQuery, state: FSMContext):
     await bot.edit_message_caption(
         chat_id=query.from_user.id,
         message_id=query.message.message_id,
-        caption=f" {caption}\nNarxi: { (ex_key.quantity+1) * price if k_type=='extended' else  (nar_key.quantity+1) * price } so'm ",
-        reply_markup=ex_key.make_order_quantity_plus() if k_type=="extended" else nar_key.make_order_quantity_plus()
+        caption=f" {caption}\Price: { (ex_key.quantity+1) * price if k_type=='extended' else  (nar_key.quantity+1) * price } sum ",
+        reply_markup=ex_key.make_order_quantity_plus() if k_type=="extended" else nar_key.make_order_quantity_plus()  # changing reply_markup for showing price
         )
     
 
@@ -69,7 +73,8 @@ async def plus_handler(query:CallbackQuery, state: FSMContext):
 
 @dp.callback_query_handler(text="-",state="order")
 async def plus_handler(query:CallbackQuery, state:FSMContext):
-    if ex_key.quantity>=1:
+    logging.info(nar_key.quantity)  # for logging 
+    if ex_key.quantity>=1 or nar_key.quantity>=1:
         data = await state.get_data()   # getting all necessary information from states
         caption = data.get("description")
         price= data.get("price")
@@ -78,9 +83,12 @@ async def plus_handler(query:CallbackQuery, state:FSMContext):
         await bot.edit_message_caption(
         chat_id=query.from_user.id,
         message_id=query.message.message_id,
-        caption=f" {caption}\nNarxi: { (ex_key.quantity-1) * price if k_type=='extended' else  (nar_key.quantity-1) * price } so'm ",
+        caption=f" {caption}\Price: { (ex_key.quantity-1) * price if k_type=='extended' else  (nar_key.quantity-1) * price } sum ",
         reply_markup=ex_key.make_order_quantity_minus() if k_type=="extended" else nar_key.make_order_quantity_minus()
         )
+
+    else:
+        query.answer("You can not reduce quantity")
 
 
 @dp.callback_query_handler(text="order_big", state="order")
@@ -108,37 +116,43 @@ async def szie2_handler(query:CallbackQuery, state: FSMContext):
 @dp.callback_query_handler(text="order",state="order")
 async def order_handler(query:CallbackQuery, state: FSMContext):
     # await query.answer("ordered", show_alert=True)
-    data = await state.get_data()
-    price = data.get("price")
-    item_name = data.get("item_name")
-    category =data.get("category")
-    k_type = data.get("k_type")
-    quantity=ex_key.quantity if k_type=="extended" else nar_key.quantity
-    size=None
-    if k_type!="extended":
-        size = None
-    elif ex_key.size1=="✅":    
-        size = "big"
-    elif ex_key.size2=="✅":
-        size="small"
+    if nar_key.quantity>=1 or ex_key.quantity>=1 :
+        if nar_key.quantity!=0   or (ex_key.size1=="" and ex_key.size2!="") or (ex_key.size1!="" and ex_key.size2==""):
+                
+            data = await state.get_data()
+            price = data.get("price")
+            item_name = data.get("item_name")
+            category =data.get("category")
+            k_type = data.get("k_type")
+            quantity=ex_key.quantity if k_type=="extended" else nar_key.quantity
+            size=None
+            if k_type!="extended":
+                size = None
+            elif ex_key.size1=="✅":    
+                size = "big"
+            elif ex_key.size2=="✅":
+                size="small"
 
 
-    await db.add_to_basket(full_name=query.from_user.full_name,
-                            telegram_id=query.from_user.id,
-                            category=category, 
-                            item_name=item_name,
-                            item_price=price,
-                            quantity=ex_key.quantity if k_type=="extended" else nar_key.quantity,
-                            item_size=size
-                            )
+            await db.add_to_basket(full_name=query.from_user.full_name,
+                                    telegram_id=query.from_user.id,
+                                    category=category, 
+                                    item_name=item_name,
+                                    item_price=price,
+                                    quantity=ex_key.quantity if k_type=="extended" else nar_key.quantity, # turnary if for checking keyborad type 
+                                    item_size=size
+                                    )
 
-    await query.message.answer("Savatga muvaffaqiyatli qo'shildi ✅")
-    ex_key.reset() #   reseting class data 
-    nar_key.reset()  #.    reseting 
-    await query.message.delete()
-    await query.message.answer("Yana nima tanlaysiz ?", reply_markup= await def_keyboard())
-    await state.set_state("menu")
+            await query.message.answer("Successfully added to basket ✅")
+            ex_key.reset() #   reseting class data 
+            nar_key.reset()  #.    reseting 
+            await query.message.delete()
+            await query.message.answer("What else do you want to choose ?", reply_markup= await def_keyboard())
+            await state.set_state("menu")
+        else:
+            await query.message.answer("Please at least you should choose a type 🧺")
 
+    else:
+        await query.message.answer("Buy at least one in order to add to basket 🧺")
+        
     
-    
-
